之前单个组件合并到组件库例子目录里
