+++ examples/mini-program-example/src/pages/component/checkbox-group/checkbox-group.scss
