+++ examples/mini-program-example/src/pages/component/grid-view/grid-view.config.ts
