+++ examples/mini-program-example/src/pages/component/list-view/list-view.config.ts
