+++ examples/mini-program-example/src/pages/component/match-media/match-media.config.ts
