+++ examples/mini-program-example/src/pages/component/page-container/page-container.config.ts
