+++ examples/mini-program-example/src/pages/component/radio-group/radio-group.config.ts
