+++ examples/mini-program-example/src/pages/component/root-portal/root-portal.config.ts
