+++ examples/mini-program-example/src/pages/component/root-portal/root-portal.scss
@@ -1,6 +1,3 @@
-@import '@tarojs/components-mpharmony/dist/style/components/page-container';
-
-
 .overlay_black {
   background: #0d0d0d;
 }
