+++ examples/mini-program-example/src/pages/component/sticky-header/sticky-header.config.ts
