+++ examples/mini-program-example/src/pages/component/sticky-header/sticky-header.tsx
@@ -1,9 +1,9 @@
 import {View, StickyHeader, StickySection} from '@tarojs/components'
 import {useLoad} from '@tarojs/taro'
-import './index.scss'
+import './sticky-header.scss'
 import React from "react";
 
-export default function Index() {
+export default function PageView() {
   useLoad(() => {
     console.log('Page loaded.')
   })
